# Transfer tensors wholesale in Metric.many